refactor(network-activity-plugin): add RequestBody component
